cheri: fs: proc: Fabricate capabilities for /proc/kcore
When accessing /proc/kcore userland gets read-only access to kernel
memory. If this is a security problem there is nothing that CHERI
can do.

However, the caller that accesses /proc/kcore (obviously) cannot
provide a valid capability. Thus fabricate one of the access to
kernel memory.

+++ fs/proc/page.c
@@ -148,7 +148,7 @@ u64 stable_page_flags(const struct page *page)
 	const struct folio *folio;
 	struct page_snapshot ps;
 	unsigned long k;
-	unsigned long mapping;
+	uintptr_t mapping;
 	bool is_anon;
 	u64 u = 0;
